retrofit debug improvement + live data on pre taboo settings view

+++ app/src/main/java/pl/politechnika/szczesny/words_world_client/model/Achievement.java
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName;
 
 public class Achievement {
     @SerializedName("id")
-    private long id;
+    private int id;
     @SerializedName("name")
     private String name;
     @SerializedName("level")
@@ -12,15 +12,14 @@ public class Achievement {
     @SerializedName("font_awesome_icon")
     private String badgeIcon;
 
-
     public Achievement() {
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
